refactor(shims): extract shared Pages Router accessor to internal helper
Create packages/vinext/src/shims/internal/pages-router-accessor.ts
with getPagesNavigationContext() and PagesNavigationContext type.

Import from it in both navigation.ts and navigation-untracked.ts
instead of duplicating the Symbol lookup and error handling. This
prevents the string key and the accessor shape from drifting across
modules if a third consumer is added later.

router.ts still writes the global accessor directly using Symbol.for
as before.
