+++ packages/vinext/src/shims/navigation.ts
@@ -337,30 +337,15 @@ export function _registerStateAccessors(accessors: _StateAccessors): void {
 // patches in unit tests that only want the router shim.
 // ---------------------------------------------------------------------------
 
-type PagesNavigationContext = {
-  pathname: string | null;
-  searchParams: URLSearchParams;
-  params: Record<string, string | string[]> | null;
-};
+import {
+  getPagesNavigationContext as _getPagesNavigationContext,
+  type PagesNavigationContext,
+} from "./internal/pages-router-accessor.js";
 
-const PAGES_NAVIGATION_ACCESSOR_KEY = Symbol.for(
-  "vinext.navigation.pagesNavigationContextAccessor",
-);
 const PAGES_NAVIGATION_NOTIFY_KEY = Symbol.for("vinext.navigation.pagesNavigationNotify");
-type _GlobalWithPagesAccessor = typeof globalThis & {
-  [PAGES_NAVIGATION_ACCESSOR_KEY]?: () => PagesNavigationContext | null;
+type _GlobalWithPagesNotify = typeof globalThis & {
   [PAGES_NAVIGATION_NOTIFY_KEY]?: () => void;
 };
-
-function _getPagesNavigationContext(): PagesNavigationContext | null {
-  const accessor = (globalThis as _GlobalWithPagesAccessor)[PAGES_NAVIGATION_ACCESSOR_KEY];
-  if (!accessor) return null;
-  try {
-    return accessor();
-  } catch {
-    return null;
-  }
-}
 
 /**
  * Get the navigation context for the current SSR/RSC render.
@@ -1147,7 +1132,7 @@ function notifyNavigationListeners(): void {
 }
 
 if (!isServer) {
-  (globalThis as _GlobalWithPagesAccessor)[PAGES_NAVIGATION_NOTIFY_KEY] = notifyNavigationListeners;
+  (globalThis as _GlobalWithPagesNotify)[PAGES_NAVIGATION_NOTIFY_KEY] = notifyNavigationListeners;
 }
 
 // Cached URLSearchParams, pathname, etc. for referential stability
